#476; splits POST /api/redis into two routes.
