+++ api/redis/Routes.js
@@ -7,4 +7,6 @@ function stateRoutes(app) {
   app.get('/api/redis', validateAccount, require('./get.js'));
   app.get('/api/redis/logs', validateAccount, require('./getLogs.js'));
   app.post('/api/redis', validateAccount, require('./post.js'));
+  app.post('/api/redis/initialize', validateAccount,
+    require('./initialize.js'));
 }
