+++ api/redis/post.js
@@ -5,22 +5,14 @@ module.exports = self;
 
 var async = require('async');
 var _ = require('underscore');
-var path = require('path');
-var fs = require('fs');
-var spawn = require('child_process').spawn;
 
-var APIAdapter = require('../../common/APIAdapter.js');
 var configHandler = require('../../common/configHandler.js');
 
 function post(req, res) {
   var bag = {
     reqBody: req.body,
-    resBody: [],
-    apiAdapter: new APIAdapter(req.headers.authorization.split(' ')[1]),
-    params: {},
-    skipStatusChange: false,
-    component: 'redis',
-    tmpScript: '/tmp/redis.sh'
+    resBody: {},
+    component: 'redis'
   };
 
   bag.who = util.format('redis|%s', self.name);
@@ -29,24 +21,13 @@ function post(req, res) {
   async.series([
       _checkInputParams.bind(null, bag),
       _get.bind(null, bag),
-      _setProcessingFlag.bind(null, bag),
-      _getReleaseVersion.bind(null, bag),
-      _generateInitializeEnvs.bind(null, bag),
-      _generateInitializeScript.bind(null, bag),
-      _writeScriptToFile.bind(null, bag),
-      _initializeRedis.bind(null, bag),
-      _postSystemIntegration.bind(null, bag),
       _post.bind(null, bag)
     ],
     function (err) {
       logger.info(bag.who, 'Completed');
-      if (!bag.skipStatusChange)
-        _setCompleteStatus(bag, err);
-
       if (err)
         return respondWithError(res, err);
 
-      //TODO: remove tmp file
       sendJSONResponse(res, bag.resBody);
     }
   );
@@ -65,181 +46,19 @@ function _get(bag, next) {
 
   configHandler.get(bag.component,
     function (err, redis) {
-      if (err) {
-        bag.skipStatusChange = true;
+      if (err)
         return next(
           new ActErr(who, ActErr.DataNotFound,
             'Failed to get ' + bag.component, err)
         );
-      }
 
-      if (_.isEmpty(redis)) {
-        bag.skipStatusChange = true;
+      if (_.isEmpty(redis))
         return next(
           new ActErr(who, ActErr.DataNotFound,
             'No configuration in database for ' + bag.component)
         );
-      }
 
       bag.config = redis;
-      return next();
-    }
-  );
-}
-
-function _setProcessingFlag(bag, next) {
-  var who = bag.who + '|' + _setProcessingFlag.name;
-  logger.verbose(who, 'Inside');
-
-  var update = {
-    isProcessing: true,
-    isFailed: false
-  };
-
-  configHandler.put(bag.component, update,
-    function (err) {
-      if (err)
-        return next(
-          new ActErr(who, ActErr.OperationFailed,
-            'Failed to update config for ' + bag.component, err)
-        );
-
-      return next();
-    }
-  );
-}
-
-function _getReleaseVersion(bag, next) {
-  var who = bag.who + '|' + _getReleaseVersion.name;
-  logger.verbose(who, 'Inside');
-
-  var query = '';
-  bag.apiAdapter.getSystemSettings(query,
-    function (err, systemSettings) {
-      if (err)
-        return next(
-          new ActErr(who, ActErr.OperationFailed,
-            'Failed to get system settings : ' + util.inspect(err))
-        );
-
-      bag.releaseVersion = systemSettings.releaseVersion;
-
-      return next();
-    }
-  );
-}
-
-function _generateInitializeEnvs(bag, next) {
-  var who = bag.who + '|' + _generateInitializeEnvs.name;
-  logger.verbose(who, 'Inside');
-
-  bag.scriptEnvs = {
-    'RUNTIME_DIR': global.config.runtimeDir,
-    'CONFIG_DIR': global.config.configDir,
-    'REDIS_HOST': global.config.admiralIP,
-    'RELEASE': bag.releaseVersion,
-    'SCRIPTS_DIR': global.config.scriptsDir,
-    'IS_INITIALIZED': bag.config.isInitialized,
-    'IS_INSTALLED': bag.config.isInstalled,
-    'REDIS_PORT': bag.config.port
-  };
-
-  return next();
-}
-
-function _generateInitializeScript(bag, next) {
-  var who = bag.who + '|' + _generateInitializeScript.name;
-  logger.verbose(who, 'Inside');
-
-  //attach header
-  var filePath = path.join(global.config.scriptsDir, '/lib/_logger.sh');
-  var headerScript = '';
-  headerScript = headerScript.concat(__applyTemplate(filePath, bag.params));
-
-  var initializeScript = headerScript;
-  filePath = path.join(global.config.scriptsDir, 'docker/installRedis.sh');
-  initializeScript = headerScript.concat(__applyTemplate(filePath, bag.params));
-
-  bag.script = initializeScript;
-  return next();
-}
-
-function _writeScriptToFile(bag, next) {
-  var who = bag.who + '|' + _writeScriptToFile.name;
-  logger.debug(who, 'Inside');
-
-  fs.writeFile(bag.tmpScript,
-    bag.script,
-    function (err) {
-      if (err) {
-        var msg = util.format('%s, Failed with err:%s', who, err);
-        return next(
-          new ActErr(
-            who, ActErr.OperationFailed, msg)
-        );
-      }
-      fs.chmodSync(bag.tmpScript, '755');
-      return next();
-    }
-  );
-}
-
-
-function _initializeRedis(bag, next) {
-  var who = bag.who + '|' + _initializeRedis.name;
-  logger.verbose(who, 'Inside');
-
-  var exec = spawn('/bin/bash',
-    ['-c', bag.tmpScript],
-    {
-      env: bag.scriptEnvs
-    }
-  );
-
-  exec.stdout.on('data',
-    function (data)  {
-      console.log(data.toString());
-    }
-  );
-
-  exec.stderr.on('data',
-    function (data)  {
-      console.log(data.toString());
-    }
-  );
-
-  exec.on('close',
-    function (exitCode)  {
-      if (exitCode > 0)
-        return next(
-          new ActErr(who, ActErr.OperationFailed,
-            'Script returned code: ' + exitCode)
-        );
-      return next();
-    }
-  );
-}
-
-function _postSystemIntegration(bag, next) {
-  var who = bag.who + '|' + _postSystemIntegration.name;
-  logger.verbose(who, 'Inside');
-
-  var postObject = {
-    name: 'redis',
-    masterName: 'url',
-    data: {
-      url: bag.scriptEnvs.REDIS_HOST + ':' + bag.config.port
-    }
-  };
-
-  bag.apiAdapter.postSystemIntegration(postObject,
-    function (err) {
-      if (err)
-        return next(
-          new ActErr(who, ActErr.OperationFailed,
-            'Failed to create system integration: ' + util.inspect(err))
-        );
-
       return next();
     }
   );
@@ -249,10 +68,12 @@ function _post(bag, next) {
   var who = bag.who + '|' + _post.name;
   logger.verbose(who, 'Inside');
 
-  var update = {
-    isInstalled: true,
-    isInitialized: true
-  };
+  var update = {};
+
+  if (_.has(bag.reqBody, 'address'))
+    update.address = bag.reqBody.address;
+  if (_.has(bag.reqBody, 'port'))
+    update.port = bag.reqBody.port;
 
   configHandler.put(bag.component, update,
     function (err) {
@@ -265,32 +86,4 @@ function _post(bag, next) {
       return next();
     }
   );
-}
-
-function _setCompleteStatus(bag, err) {
-  var who = bag.who + '|' + _setCompleteStatus.name;
-  logger.verbose(who, 'Inside');
-
-  var update = {
-    isProcessing: false
-  };
-  if (err)
-    update.isFailed = true;
-  else
-    update.isFailed = false;
-
-  configHandler.put(bag.component, update,
-    function (err) {
-      if (err)
-        logger.warn(err);
-    }
-  );
-}
-
-//local function to apply vars to template
-function __applyTemplate(filePath, dataObj) {
-  var fileContent = fs.readFileSync(filePath).toString();
-  var template = _.template(fileContent);
-
-  return template({obj: dataObj});
 }
